+++ src/test/java/org/biouno/unochoice/BaseUiTest.java
@@ -38,6 +38,7 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 import java.text.MessageFormat;
@@ -189,5 +190,9 @@ public abstract class BaseUiTest {
                         return false;
                     }
                 });
+    }
+
+    protected void waitLoadingMessage() {
+        wait.until(ExpectedConditions.invisibilityOfElementLocated(By.cssSelector(".jenkins-spinner")));
     }
 }